backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.35.0
pydantic==2.11.9
pandas==2.3.2
python-multipart==0.0.20
httpx==0.28.1
pytest==8.4.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==45.0.7
email-validator==2.2.0
